+++ src/test/__mocks__/common/hooks/useServicesContext.mock.ts
@@ -51,6 +51,12 @@ export const recordGeneratorService = {
   generate: jest.fn(),
 } as IRecordGeneratorService;
 
+export const schemaGeneratorService = {
+  init: jest.fn(),
+  get: jest.fn(),
+  generate: jest.fn(),
+} as ISchemaGeneratorService;
+
 jest.mock('@common/hooks/useServicesContext.ts', () => ({
   useServicesContext: () => ({
     userValuesService,
@@ -62,5 +68,6 @@ jest.mock('@common/hooks/useServicesContext.ts', () => ({
     schemaCreatorService,
     recordGeneratorServiceLegacy,
     recordGeneratorService,
+    schemaGeneratorService,
   }),
 }));